#195 Refactor: combine templates folder and components folder
