+++ javascript/big-practice/src/scripts/components/noteTemplate.js
