+++ javascript/big-practice/src/scripts/components/searchTemplate.js
