+++ javascript/big-practice/src/scripts/views/listNoteView.js
@@ -6,8 +6,8 @@ import STORAGE_KEYS from '../constants/storageKeys';
 import renderConfirmPopup from '../utils/confirmPopup';
 import { POPUP_MESSAGE } from '../constants/message';
 import LocalStorage from '../utils/localStorage';
-import formTemplate from '../templates/formTemplate';
-import noteTemplate from '../templates/noteTemplate';
+import formTemplate from '../components/formTemplate';
+import noteTemplate from '../components/noteTemplate';
 import navigatePage from '../utils/navigatePage';
 import HeaderView from './headerView';
 import { renderPopupError } from '../utils/handleError';
